app2_client_without_csm/requirements.txt
fastapi
uvicorn
httpx

requests
google-auth